Day 11/customer-support-ai/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
rank-bm25==0.2.2
requests==2.32.3
python-dotenv==1.0.1
